+++ sidebar.d.ts
@@ -1,0 +1,11 @@
+export interface MSidebarLayoutElement extends HTMLElement {
+  open(): void;
+  close(): void;
+  toggle(): void;
+}
+declare global {
+  interface HTMLElementTagNameMap {
+    "m-sidebar": HTMLElement;
+    "m-sidebar-layout": MSidebarLayoutElement;
+  }
+}