+++ types/react.d.ts
@@ -51,6 +51,8 @@ declare module 'react' {
 
       // components
       'm-segmented': MicaElement
+      'm-sidebar': MicaElement
+      'm-sidebar-layout': MicaElement<{ variant?: 'docked' | 'inset'; collapse?: 'icon' }>
       'm-header': MicaElement<{ align?: 'start' | 'end' }>
       'm-stepper': MicaElement
       'm-tabs': MicaElement<{ variant?: 'underline' }>